Split main.cpp into headers and cpp files.
