+++ .replit
@@ -1,2 +1,2 @@
 language = "cpp"
-run = "clang++ -std=c++17 -Weverything -Wno-exit-time-destructors -Wno-padded -Wno-c++98-compat -Wno-missing-prototypes -Wno-deprecated main.cpp && ./a.out"
+run = "clang++ -std=c++17 -Weverything -Wno-exit-time-destructors -Wno-padded -Wno-c++98-compat -Wno-missing-prototypes -Wno-deprecated main.cpp Note.cpp GuitarString.cpp Chord.cpp Fretboard.cpp ChordProgression.cpp && ./a.out"
