+++ Defines.h
@@ -1,0 +1,13 @@
+/******************************************************************
+    Defines.h
+
+    Global constants defines and enumerations
+******************************************************************/
+#pragma once
+#define NUM_STRINGS 6
+#define NUM_FRETS 24
+#define NUM_KEYS 12
+#define NUM_CHORDS 3
+
+// enumeration for fingers on the guitar fretboard
+enum Finger { thumb, index, middle, ring, pinky, none };
